refactor(tests): extract shared mock patterns into test_utils module
Move StubClient and TableStub classes from test_memory_recall_hook.py into
a shared tests/test_utils.py so test_memory_server.py and future test
modules can reuse them without duplication. Also remove redundant
module-level stubs from test_memory_recall_hook.py that are already
handled by conftest.py, and add create_client to conftest.py's supabase
stub to support this cleanup.

Closes #860
